split out layout/dialogs from app.js
